# Add caching to IFont2Font