Adding caching to font creation as it can be quite costly. Showing the difference it makes by creating benchmark for `AutoSizeColumn`.

## NPOI.Benchmarks.AutoSizeColumnBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |AutoSizeColumn|1.329 s|0.0156 s|673.13 MB|
| **New** |	| **144.3 ms (-89%)** | **1.94 ms** | **130.58 MB (-81%)** |