@lahma 2023-09-06
> The caching result is really impressive! @lahma, maybe ceiling `FontHeightInPoints` will be a better approach?

I'm hoping that generally the value is always an integer so shouldn't matter that much. Just wanted to make sure nobody creates a font-loader bomb with an excel that has different font size in every cell and just by decimals.