@Sappharad 2023-09-06
Thank you for improving the performance so soon after I commented on it yesterday, the benchmarks you posted look great! I rolled back to 2.5.6 but look forward to upgrading to 2.7.0 when it's out.