@JimBobSquarePants 2023-09-08
This is the correct approach. ClosedXML has introduced incorrect measurements by using custom measurements based upon the advance of individual glyphs. 

I've worked hard to ensure that text measurement is fast and accurate (_proper bidi layout and shaping is incredibly complicated_) but would always recommend a font file is primed for consumption.